refactor(remove-pandacss): 내 프로필 페이지 복구
